src/reflex_conversion/requirements.txt

reflex>=0.3.4
dataverse-api-client==0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
